StorageDirectory breakdown example
