+++ src/_versions.ts
@@ -11,8 +11,8 @@ export interface TsAppVersion {
 export const versions: TsAppVersion = {
     version: '0.6.0',
     name: 'nachet-frontend',
-    versionDate: '2024-04-09T22:31:42.159Z',
-    gitCommitHash: '6b3e2ef',
-    versionLong: '0.6.0-6b3e2ef',
+    versionDate: '2024-04-10T00:39:01.552Z',
+    gitCommitHash: 'd288002',
+    versionLong: '0.6.0-d288002',
 };
 export default versions;
