+++ src/components/body/directory_list/StorageDirectoryContainer.tsx
@@ -1,0 +1,39 @@
+import StorageDirectoryView from "./StorageDirectoryView";
+
+export interface params {
+  azureStorageDir: any;
+  curDir: string;
+  setCurDir: React.Dispatch<React.SetStateAction<string>>;
+  setCreateDirectoryOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setDelDirectoryOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  handleDirChange: (dir: string) => void;
+}
+
+const StorageDirectoryContainer: React.FC<params> = (props): JSX.Element => {
+  const handleDelete = (folder: string): void => {
+    props.handleDirChange(folder);
+    props.setDelDirectoryOpen(true);
+  };
+  const handleSelect = (folder: string): void => {
+    if (folder === props.curDir) {
+      props.handleDirChange("General");
+    } else {
+      props.handleDirChange(folder);
+    }
+  };
+  const handleCreateDirectory = (): void => {
+    props.setCreateDirectoryOpen(true);
+    props.setCurDir("");
+  };
+
+  return (
+    <StorageDirectoryView
+      {...props}
+      handleSelect={handleSelect}
+      handleDelete={handleDelete}
+      handleCreateDirectory={handleCreateDirectory}
+    />
+  );
+};
+
+export default StorageDirectoryContainer;
